+++ spec/models/person_spec.rb
@@ -4,73 +4,75 @@
 #   licensed under the Affero General Public License version 3 or later.  See
 #   the COPYRIGHT file.
 
-describe Person, :type => :model do
+describe Person, type: :model do
   before do
     @user = bob
     @person = FactoryBot.create(:person)
   end
 
-  it 'always has a profile' do
+  it "always has a profile" do
     expect(Person.new.profile).not_to be_nil
   end
 
-  it 'does not save automatically' do
+  it "does not save automatically" do
     expect(Person.new.persisted?).to be false
     expect(Person.new.profile.persisted?).to be false
   end
 
-  context 'scopes' do
-    describe '.for_json' do
-      it 'does not select public keys' do
+  context "scopes" do
+    describe ".for_json" do
+      it "does not select public keys" do
         expect {
           Person.for_json.first.serialized_public_key
         }.to raise_error ActiveModel::MissingAttributeError
       end
     end
 
-    describe '.local' do
-      it 'returns only local people' do
+    describe ".local" do
+      it "returns only local people" do
         Person.local =~ [@person]
       end
     end
 
-    describe '.remote' do
-      it 'returns only local people' do
+    describe ".remote" do
+      it "returns only local people" do
         Person.remote =~ [@user.person]
       end
     end
 
-    describe '.find_person_from_guid_or_username' do
-      it 'searchs for a person if id is passed' do
-        expect(Person.find_from_guid_or_username(:id => @person.guid).id).to eq(@person.id)
+    describe ".find_person_from_guid_or_username" do
+      it "searchs for a person if id is passed" do
+        expect(Person.find_from_guid_or_username(id: @person.guid).id).to eq(@person.id)
       end
 
-      it 'searchs a person from a user if username is passed' do
-        expect(Person.find_from_guid_or_username(:username => @user.username).id).to eq(@user.person.id)
+      it "searchs a person from a user if username is passed" do
+        expect(Person.find_from_guid_or_username(username: @user.username).id).to eq(@user.person.id)
       end
 
-      it 'throws active record not found exceptions if no person is found via id' do
-        expect{
-          Person.find_from_guid_or_username(:id => "2d13123")
+      it "throws active record not found exceptions if no person is found via id" do
+        expect {
+          Person.find_from_guid_or_username(id: "2d13123")
         }.to raise_error ActiveRecord::RecordNotFound
       end
 
-      it 'throws active record not found exceptions if no person is found via username' do
-        expect{
-          Person.find_from_guid_or_username(:username => 'michael_jackson')
+      it "throws active record not found exceptions if no person is found via username" do
+        expect {
+          Person.find_from_guid_or_username(username: "michael_jackson")
         }.to raise_error ActiveRecord::RecordNotFound
       end
     end
 
-    describe '.all_from_aspects' do
+    describe ".all_from_aspects" do
       it "pulls back the right people given all a user's aspects" do
         aspect_ids = bob.aspects.map(&:id)
-        expect(Person.all_from_aspects(aspect_ids, bob).map(&:id)).to match_array(bob.contacts.includes(:person).map{|c| c.person.id})
+        expect(Person.all_from_aspects(aspect_ids, bob).map(&:id))
+          .to match_array(bob.contacts.includes(:person).map {|c| c.person.id })
       end
 
       it "pulls back the right people given a subset of aspects" do
         aspect_ids = bob.aspects.first.id
-        expect(Person.all_from_aspects(aspect_ids, bob).map(&:id)).to match_array(bob.aspects.first.contacts.includes(:person).map{|c| c.person.id})
+        expect(Person.all_from_aspects(aspect_ids, bob).map(&:id))
+          .to match_array(bob.aspects.first.contacts.includes(:person).map {|c| c.person.id })
       end
 
       it "respects aspects given a user" do
@@ -80,13 +82,14 @@ describe Person, :type => :model do
     end
 
     describe ".who_have_reshared a user's posts" do
-      it 'pulls back users who reshared the status message of a user' do
-        sm = FactoryBot.create(:status_message, :author => alice.person, :public => true)
-        reshare = FactoryBot.create(:reshare, :root => sm)
+      it "pulls back users who reshared the status message of a user" do
+        sm = FactoryBot.create(:status_message, author: alice.person, public: true)
+        reshare = FactoryBot.create(:reshare, root: sm)
         expect(Person.who_have_reshared_a_users_posts(alice)).to eq([reshare.author])
       end
     end
 
+    # rubocop:disable Rails/DynamicFindBy
     describe ".find_by_substring" do
       it "returns \"none\" when the substring is less than 1 non-space character" do
         expect(Person.find_by_substring("R")).to eq(Person.none)
@@ -136,6 +139,7 @@ describe Person, :type => :model do
         expect(Person.allowed_to_be_mentioned_in_a_comment_to(status_bob).ids).to match_array(Person.ids)
       end
     end
+    # rubocop:enable Rails/DynamicFindBy
 
     describe ".sort_for_mention_suggestion" do
       let(:status_message) { FactoryBot.create(:status_message) }
@@ -203,7 +207,7 @@ describe Person, :type => :model do
   describe "delegating" do
     it "delegates last_name to the profile" do
       expect(@person.last_name).to eq(@person.profile.last_name)
-      @person.profile.update_attributes(:last_name => "Heathers")
+      @person.profile.update(last_name: "Heathers")
       expect(@person.reload.last_name).to eq("Heathers")
     end
   end
@@ -262,74 +266,80 @@ describe Person, :type => :model do
     end
   end
 
-  describe '#diaspora_handle' do
-    context 'local people' do
-      it 'uses the pod config url to set the diaspora_handle' do
-        new_person = User.build(:username => "foo123", :email => "foo123@example.com", :password => "password", :password_confirmation => "password").person
+  describe "#diaspora_handle" do
+    context "local people" do
+      it "uses the pod config url to set the diaspora_handle" do
+        new_person = User.build(username: "foo123", email: "foo123@example.com",
+          password: "password", password_confirmation: "password").person
         expect(new_person.diaspora_handle).to eq("foo123#{User.diaspora_id_host}")
       end
 
-      it 'does not include www if it is set in app config' do
-        allow(AppConfig).to receive(:pod_uri).and_return(Addressable::URI.parse('https://www.foobar.com/'))
-        new_person = User.build(:username => "foo123", :email => "foo123@example.com", :password => "password", :password_confirmation => "password").person
+      it "does not include www if it is set in app config" do
+        allow(AppConfig).to receive(:pod_uri).and_return(Addressable::URI.parse("https://www.foobar.com/"))
+        new_person = User.build(username: "foo123", email: "foo123@example.com", password: "password",
+          password_confirmation: "password").person
         expect(new_person.diaspora_handle).to eq("foo123@foobar.com")
       end
     end
 
-    context 'remote people' do
-      it 'stores the diaspora_handle in the database' do
+    context "remote people" do
+      it "stores the diaspora_handle in the database" do
         expect(@person.diaspora_handle.include?(AppConfig.pod_uri.host)).to be false
       end
     end
 
-    describe 'validation' do
-      it 'is unique' do
-        person_two = FactoryBot.build(:person, :diaspora_handle => @person.diaspora_handle)
+    describe "validation" do
+      it "is unique" do
+        person_two = FactoryBot.build(:person, diaspora_handle: @person.diaspora_handle)
         expect(person_two).not_to be_valid
       end
 
-      it 'is case insensitive' do
-        person_two = FactoryBot.build(:person, :diaspora_handle => @person.diaspora_handle.upcase)
+      it "is case insensitive" do
+        person_two = FactoryBot.build(:person, diaspora_handle: @person.diaspora_handle.upcase)
         expect(person_two).not_to be_valid
       end
     end
   end
 
-  context '.name_from_attrs' do
+  context ".name_from_attrs" do
     before do
       @person = alice.person
       @profile = @person.profile
     end
 
-    context 'with only first name' do
-      it 'should return their first name for name' do
-        expect(Person.name_from_attrs(@profile.first_name, nil, @profile.diaspora_handle)).to eq(@profile.first_name.strip)
+    context "with only first name" do
+      it "should return their first name for name" do
+        expect(Person.name_from_attrs(@profile.first_name, nil, @profile.diaspora_handle))
+          .to eq(@profile.first_name.strip)
       end
     end
 
-    context 'with only last name' do
-      it 'should return their last name for name' do
-        expect(Person.name_from_attrs(nil, @profile.last_name, @profile.diaspora_handle)).to eq(@profile.last_name.strip)
+    context "with only last name" do
+      it "should return their last name for name" do
+        expect(Person.name_from_attrs(nil, @profile.last_name, @profile.diaspora_handle))
+          .to eq(@profile.last_name.strip)
       end
     end
 
-    context 'with both first and last name' do
-      it 'should return their composed name for name' do
-        expect(Person.name_from_attrs(@profile.first_name, @profile.last_name, @profile.diaspora_handle)).to eq("#{@profile.first_name.strip} #{@profile.last_name.strip}")
+    context "with both first and last name" do
+      it "should return their composed name for name" do
+        expect(Person.name_from_attrs(@profile.first_name, @profile.last_name, @profile.diaspora_handle))
+          .to eq("#{@profile.first_name.strip} #{@profile.last_name.strip}")
       end
     end
 
-    context 'without first nor last name' do
-      it 'should display their diaspora handle' do
+    context "without first nor last name" do
+      it "should display their diaspora handle" do
         expect(Person.name_from_attrs(nil, nil, @profile.diaspora_handle)).to eq(@profile.diaspora_handle)
       end
     end
   end
 
-  describe '#name' do
-    it 'calls Person.name_from_attrs' do
+  describe "#name" do
+    it "calls Person.name_from_attrs" do
       profile = alice.person.profile
-      expect(Person).to receive(:name_from_attrs).with(profile.first_name, profile.last_name, profile.person.diaspora_handle)
+      expect(Person).to receive(:name_from_attrs)
+        .with(profile.first_name, profile.last_name, profile.person.diaspora_handle)
       alice.name
     end
 
@@ -341,8 +351,8 @@ describe Person, :type => :model do
     end
   end
 
-  it '#owns? posts' do
-    person_message = FactoryBot.create(:status_message, :author => @person)
+  it "#owns? posts" do
+    person_message = FactoryBot.create(:status_message, author: @person)
     person_two = FactoryBot.create(:person)
 
     expect(@person.owns?(person_message)).to be true
@@ -352,41 +362,41 @@ describe Person, :type => :model do
   describe "disconnecting" do
     before do
       @user2 = FactoryBot.create(:user)
-      @aspect = @user.aspects.create(:name => "Dudes")
-      @aspect2 = @user2.aspects.create(:name => "Abscence of Babes")
+      @aspect = @user.aspects.create(name: "Dudes")
+      @aspect2 = @user2.aspects.create(name: "Abscence of Babes")
     end
-    it 'should not delete an orphaned contact' do
-      @user.contacts.create(:person => @person, :aspects => [@aspect])
+    it "should not delete an orphaned contact" do
+      @user.contacts.create(person: @person, aspects: [@aspect])
 
       expect { @user.disconnect(@user.contact_for(@person)) }.not_to change(Person, :count)
     end
 
-    it 'should not delete an un-orphaned contact' do
-      @user.contacts.create(:person => @person, :aspects => [@aspect])
-      @user2.contacts.create(:person => @person, :aspects => [@aspect2])
+    it "should not delete an un-orphaned contact" do
+      @user.contacts.create(person: @person, aspects: [@aspect])
+      @user2.contacts.create(person: @person, aspects: [@aspect2])
 
       expect { @user.disconnect(@user.contact_for(@person)) }.not_to change(Person, :count)
     end
   end
 
   describe "#first_name" do
-    it 'returns username if first_name is not present in profile' do
-      alice.person.profile.update_attributes(:first_name => "")
+    it "returns username if first_name is not present in profile" do
+      alice.person.profile.update(first_name: "")
       expect(alice.person.first_name).to eq(alice.username)
     end
 
-    it 'returns first words in first_name if first_name is present' do
-      alice.person.profile.update_attributes(:first_name => "First Mid Last")
+    it "returns first words in first_name if first_name is present" do
+      alice.person.profile.update(first_name: "First Mid Last")
       expect(alice.person.first_name).to eq("First Mid")
     end
 
-    it 'returns first word in first_name if first_name is present' do
-      alice.person.profile.update_attributes(:first_name => "Alice")
+    it "returns first word in first_name if first_name is present" do
+      alice.person.profile.update(first_name: "Alice")
       expect(alice.person.first_name).to eq("Alice")
     end
   end
 
-  describe '.search' do
+  describe ".search" do
     before do
       Person.delete_all
       @user = FactoryBot.create(:user_with_aspect)
@@ -434,7 +444,7 @@ describe Person, :type => :model do
       @closed_account.reload
     end
 
-    it 'orders results by last name' do
+    it "orders results by last name" do
       @robert_grimm.profile.first_name = "AAA"
       @robert_grimm.profile.save!
 
@@ -448,20 +458,20 @@ describe Person, :type => :model do
       @casey_grippi.profile.save!
 
       people = Person.search("AAA", @user)
-      expect(people.map { |p| p.name }).to eq([@yevgeniy_dodis, @robert_grimm, @casey_grippi, @eugene_weinstein].map { |p| p.name })
+      expect(people.map(&:name)).to eq([@yevgeniy_dodis, @robert_grimm, @casey_grippi, @eugene_weinstein].map(&:name))
     end
 
-    it 'returns nothing on an empty query' do
+    it "returns nothing on an empty query" do
       people = Person.search("", @user)
       expect(people).to be_empty
     end
 
-    it 'returns nothing on a one-character query' do
+    it "returns nothing on a one-character query" do
       people = Person.search("i", @user)
       expect(people).to be_empty
     end
 
-    it 'returns results for partial names' do
+    it "returns results for partial names" do
       people = Person.search("Eug", @user)
       expect(people.count).to eq(1)
       expect(people.first).to eq(@eugene_weinstein)
@@ -476,7 +486,7 @@ describe Person, :type => :model do
       expect(people.second).to eq(@casey_grippi)
     end
 
-    it 'returns results for full names' do
+    it "returns results for full names" do
       people = Person.search("Casey Grippi", @user)
       expect(people.count).to eq(1)
       expect(people.first).to eq(@casey_grippi)
@@ -499,7 +509,7 @@ describe Person, :type => :model do
       expect(people.first).to eq(@invisible_person)
     end
 
-    it 'returns results for Diaspora handles' do
+    it "returns results for Diaspora handles" do
       people = Person.search(@robert_grimm.diaspora_handle, @user)
       expect(people).to eq([@robert_grimm])
     end
@@ -517,10 +527,10 @@ describe Person, :type => :model do
       @casey_grippi.profile.first_name = "AAA"
       @casey_grippi.profile.save!
 
-      @user.contacts.create(:person => @casey_grippi, :aspects => [@user.aspects.first])
+      @user.contacts.create(person: @casey_grippi, aspects: [@user.aspects.first])
 
       people = Person.search("AAA", @user)
-      expect(people.map { |p| p.name }).to eq([@casey_grippi, @yevgeniy_dodis, @robert_grimm, @eugene_weinstein].map { |p| p.name })
+      expect(people.map(&:name)).to eq([@casey_grippi, @yevgeniy_dodis, @robert_grimm, @eugene_weinstein].map(&:name))
     end
 
     context "only contacts" do
@@ -591,39 +601,39 @@ describe Person, :type => :model do
     end
 
     it "handles broken keys and returns nil" do
-      @person.update_attributes(serialized_public_key: "broken")
+      @person.update(serialized_public_key: "broken")
       expect(@person.public_key).to be_nil
     end
   end
 
-  context 'people finders for webfinger' do
+  context "people finders for webfinger" do
     let(:user) { FactoryBot.create(:user) }
     let(:person) { FactoryBot.create(:person) }
 
-    describe '.by_account_identifier' do
-      it 'should find a local users person' do
+    describe ".by_account_identifier" do
+      it "should find a local users person" do
         p = Person.by_account_identifier(user.diaspora_handle)
         expect(p).to eq(user.person)
       end
 
-      it 'should find remote users person' do
+      it "should find remote users person" do
         p = Person.by_account_identifier(person.diaspora_handle)
         expect(p).to eq(person)
       end
 
-      it 'should downcase and strip the diaspora_handle' do
+      it "should downcase and strip the diaspora_handle" do
         dh_upper = "    " + user.diaspora_handle.upcase + "   "
         expect(Person.by_account_identifier(dh_upper)).to eq(user.person)
       end
 
       it "finds a local person with a mixed-case username" do
-        user = FactoryBot.create(:user, :username => "SaMaNtHa")
+        user = FactoryBot.create(:user, username: "SaMaNtHa")
         person = Person.by_account_identifier(user.person.diaspora_handle)
         expect(person).to eq(user.person)
       end
 
       it "is case insensitive" do
-        user1 = FactoryBot.create(:user, :username => "SaMaNtHa")
+        user1 = FactoryBot.create(:user, username: "SaMaNtHa")
         person = Person.by_account_identifier(user1.person.diaspora_handle.upcase)
         expect(person).to eq(user1.person)
       end
@@ -643,20 +653,20 @@ describe Person, :type => :model do
     end
   end
 
-  describe '#has_photos?' do
-    it 'returns false if the user has no photos' do
+  describe "#has_photos?" do
+    it "returns false if the user has no photos" do
       expect(alice.person.has_photos?).to be false
     end
 
-    it 'returns true if the user has photos' do
-      alice.post(:photo, :user_file => uploaded_photo, :to => alice.aspects.first.id)
+    it "returns true if the user has photos" do
+      alice.post(:photo, user_file: uploaded_photo, to: alice.aspects.first.id)
 
       expect(alice.person.has_photos?).to be true
     end
   end
 
-  describe '#as_json' do
-    it 'returns a hash representation of a person' do
+  describe "#as_json" do
+    it "returns a hash representation of a person" do
       expect(@person.as_json).to eq(
         id:     @person.id,
         guid:   @person.guid,
@@ -666,15 +676,15 @@ describe Person, :type => :model do
         url:    Rails.application.routes.url_helpers.person_path(@person)
       )
     end
-    it 'return tags if asked' do
-      expect(@person.as_json(:includes => "tags")).
-        to eq(@person.as_json.merge(:tags => @person.profile.tags.map { |t| "##{t.name}" }))
+    it "return tags if asked" do
+      expect(@person.as_json(includes: "tags"))
+        .to eq(@person.as_json.merge(tags: @person.profile.tags.map {|t| "##{t.name}" }))
     end
   end
 
-  describe '.community_spotlight' do
+  describe ".community_spotlight" do
     describe "when the pod owner hasn't set up any community spotlight members" do
-      it 'returns people with the community spotlight role' do
+      it "returns people with the community spotlight role" do
         Role.add_spotlight(bob.person)
         expect(Person.community_spotlight).to be_present
       end
@@ -685,8 +695,8 @@ describe Person, :type => :model do
     end
   end
 
-  describe '#lock_access!' do
-    it 'sets the closed_account flag' do
+  describe "#lock_access!" do
+    it "sets the closed_account flag" do
       @person.lock_access!
       expect(@person.reload.closed_account).to be true
     end
@@ -697,7 +707,7 @@ describe Person, :type => :model do
       @person = FactoryBot.build(:person)
     end
 
-    it 'calls Profile#tombstone!' do
+    it "calls Profile#tombstone!" do
       expect(@person.profile).to receive(:tombstone!)
       @person.clear_profile!
     end